requirements.txt
bintrees==2.0.7
boto3==1.4.4
botocore==1.5.18
docutils==0.13.1
futures==3.0.5
gdax==1.0.6
jmespath==0.9.3
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
s3transfer==0.1.10
six==1.10.0
websocket-client==0.40.0
